hostelmanager/hostelmanager.requirements.txt
insbackports-abc==0.4
click==6.6
defusedxml==0.4.1
Django==1.9.4
django-allauth==0.25.2
django-appconf==1.0.1
django-common-helpers==0.8.0
django-cron==0.4.6
django-filter==0.13.0
django-geoposition==0.2.2
django-image-cropping==1.0.3
django-templated-email==0.4.9
djangorestframework==3.3.3
djangorestframework-jwt==1.7.2
djangorestframework-xml==1.3.0
easy-thumbnails==2.3
Jinja2==2.8
livereload==2.4.1
Markdown==2.6.6 
MarkupSafe==0.23
MySQL-python==1.2.3
mysqlclient==1.3.7
oauthlib==1.0.3
Pillow==3.1.1
Pmw==2.0.0
preppy==2.3.5
PyJWT==1.4.0
PyMySQL==0.7.2
pyRXP==2.1.1
python3-openid==3.0.10
PyYAML==3.11
reportlab==3.3.3
requests==2.9.1
requests-oauthlib==0.6.1
rlextra==3.3.6
six==1.10.0
StringGenerator==0.1.9
tornado==4.3 